refactored to move code in utils and updated redundant docstrings
